+++ packages/brightcove-video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.10.1"></a>
+## [1.10.1](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@1.10.0...@times-components/brightcove-video@1.10.1) (2018-01-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
 <a name="1.10.0"></a>
 # [1.10.0](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@1.9.5...@times-components/brightcove-video@1.10.0) (2018-01-04)